Rename file to match class name
