+++ lib/derailed_benchmarks/stats_for_file.rb
